% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGMissingDataR-package.R
\docType{package}
\name{CGMissingDataR-package}
\alias{CGMissingDataR}
\alias{CGMissingDataR-package}
\title{CGMissingDataR: Missingness Benchmark for Continuous Glucose Monitoring Data}
\description{
Evaluates predictive performance under feature-level missingness in repeated-measures continuous glucose monitoring-like data. The benchmark injects missing values at user-specified rates, imputes incomplete feature matrices using an iterative chained-equations approach inspired by multivariate imputation by chained equations (MICE; Azur et al. (2011) \doi{10.1002/mpr.329}), fits Random Forest regression models (Breiman (2001) \doi{10.1023/A:1010933404324}) and k-nearest-neighbor regression models (Zhang (2016) \doi{10.21037/atm.2016.03.37}), and reports mean absolute percentage error and R-squared across missingness rates.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://zhanglabuky.github.io/CGMissingDataR/}
  \item \url{https://github.com/ZhangLabUKY/CGMissingDataR}
  \item Report bugs at \url{https://github.com/ZhangLabUKY/CGMissingDataR/issues}
}

}
\author{
\strong{Maintainer}: Shubh Saraswat \email{shubh.saraswat00@gmail.com} (\href{https://orcid.org/0009-0009-2359-1484}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Hasin Shahed Shad \email{hasin.shad@uky.edu}
  \item Xiaohua Douglas Zhang \email{douglas.zhang@uky.edu} (\href{https://orcid.org/0000-0002-2486-7931}{ORCID})
}

}
\keyword{internal}
